back_end/requirements.txt
Django==5.0.4
django-environ==0.11.2
python-dotenv==1.0.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
ccxt==4.3.28
GoogleNews==1.6.14
pandas==2.2.2
sqlalchemy==2.0.30

# POSTGRES
#If running on Linux, the below commands must be entered first
# sudo apt-get install python-pip
# sudo apt-get update
# sudo apt-get install libpq-dev python2-dev python2 python-dev-is-python3
psycopg2==2.9.9